+++ anubis-management-api/anubis/tests/test_policies.py
@@ -26,6 +26,14 @@ def test_policies(test_db):
     assert tenant_id
 
     response = client.post(
+        "/v1/tenants/" + tenant_id + "/service_paths",
+        json={"path": "/foobar"}
+    )
+    assert response.status_code == 201
+    service_path_id = response.headers["Service-Path-ID"]
+    assert service_path_id
+
+    response = client.post(
         "/v1/policies/",
         headers={
             "fiware-service": "test",
@@ -67,11 +75,11 @@ def test_policies(test_db):
     policy_id = response.headers["Policy-ID"]
     assert policy_id
 
+    # test policy creation without servicePath
     response = client.post(
         "/v1/policies/",
         headers={
-            "fiware-service": "test",
-            "fiware-servicepath": "/"},
+            "fiware-service": "test" },
         json={
             "access_to": "resource",
             "resource_type": "entity",
@@ -81,6 +89,18 @@ def test_policies(test_db):
     policy_id = response.headers["Policy-ID"]
     assert policy_id
 
+    response = client.post(
+        "/v1/policies/",
+        headers={
+            "fiware-service": "test",
+            "fiware-servicepath": "/foobar"},
+        json={
+            "access_to": "resource",
+            "resource_type": "entity",
+            "mode": ["acl:Write"],
+            "agent": ["acl:agent:test"]})
+    assert response.status_code == 201
+
     response = client.get(
         "/v1/policies/",
         headers={
@@ -89,6 +109,23 @@ def test_policies(test_db):
     assert response.status_code == 200
     body = response.json()
     assert len(body) == 8
+
+    response = client.get(
+        "/v1/policies/",
+        headers={
+            "fiware-service": "test",
+            "fiware-servicepath": "/foobar"})
+    assert response.status_code == 200
+    body = response.json()
+    assert len(body) == 1
+
+    response = client.get(
+        "/v1/policies/",
+        headers={
+            "fiware-service": "test"})
+    assert response.status_code == 200
+    body = response.json()
+    assert len(body) == 9
 
     response = client.get(
         "/v1/policies/?agent_type=acl:agent",
@@ -118,7 +155,7 @@ def test_policies(test_db):
     assert len(body) == 7
 
     response = client.get(
-        "/v1/policies/?resource_type=entity&&agent_type=acl:agent",
+        "/v1/policies/?resource_type=entity&agent_type=acl:agent",
         headers={
             "fiware-service": "test",
             "fiware-servicepath": "/"})
@@ -127,7 +164,7 @@ def test_policies(test_db):
     assert len(body) == 4
 
     response = client.get(
-        "/v1/policies/?resource_type=entity&&agent_type=acl:agentGroup",
+        "/v1/policies/?resource_type=entity&agent_type=acl:agentGroup",
         headers={
             "fiware-service": "test",
             "fiware-servicepath": "/"})
@@ -136,7 +173,7 @@ def test_policies(test_db):
     assert len(body) == 1
 
     response = client.get(
-        "/v1/policies/?resource_type=entity&&agent_type=acl:agent&&resource=resource",
+        "/v1/policies/?resource_type=entity&agent_type=acl:agent&resource=resource",
         headers={
             "fiware-service": "test",
             "fiware-servicepath": "/"})
@@ -145,13 +182,31 @@ def test_policies(test_db):
     assert len(body) == 4
 
     response = client.get(
-        "/v1/policies/?resource_type=entity&&agent_type=acl:agent&&resource=resource:foobar",
+        "/v1/policies/?resource_type=entity&agent_type=acl:agent&resource=resource:foobar",
         headers={
             "fiware-service": "test",
             "fiware-servicepath": "/"})
     body = response.json()
     assert response.status_code == 200
     assert len(body) == 2
+
+    response = client.get(
+        "/v1/policies/?resource_type=entity&agent_type=acl:agent&resource=resource:foobar&mode=acl:Read",
+        headers={
+            "fiware-service": "test",
+            "fiware-servicepath": "/"})
+    body = response.json()
+    assert response.status_code == 200
+    assert len(body) == 2
+
+    response = client.get(
+        "/v1/policies/?resource_type=entity&agent=acl:agent:test",
+        headers={
+            "fiware-service": "test",
+            "fiware-servicepath": "/"})
+    body = response.json()
+    assert response.status_code == 200
+    assert len(body) == 4
 
     response = client.get(
         "/v1/policies/?resource_type=subscription",
@@ -200,6 +255,34 @@ def test_policies(test_db):
             "agent": ["acl:agent:test"]})
     assert response.status_code == 201
     policy_id = response.headers["policy-id"]
+
+
+    response = client.put(
+        "/v1/policies/" + policy_id,
+        headers={
+            "fiware-service": "Tenant1",
+            "fiware-servicepath": "/"},
+        json={
+            "access_to": "resource-2",
+            "resource_type": "entity",
+            "mode": ["acl:Read"],
+            "agent": ["acl:agent:test"]})
+    assert response.status_code == 201
+    policy_id = response.headers["policy-id"]
+
+
+    response = client.get(
+        "/v1/policies/" + policy_id,
+        headers={
+            "fiware-service": "Tenant1",
+            "fiware-servicepath": "/"})
+    body = response.json()
+    assert response.status_code == 200
+    assert body["access_to"] == "resource-2"
+    assert body["resource_type"] == "entity"
+    assert body["mode"] == ["acl:Read"]
+    assert body["agent"] == ["acl:agent:test"]
+
 
     response = client.delete(
         "/v1/policies/" + policy_id,